requirements.txt
hyperopt==0.2.7
matplotlib==3.6.2
numpy==1.22.3
opencv_python==4.5.4.58
pandas==1.3.3
Pillow==9.5.0
scipy==1.9.1
